Deleted fixtures file
